tests/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
dill==0.3.2
distro==1.5.0
genie==20.6.1
genie.libs.clean==20.6.1
genie.libs.conf==20.6
genie.libs.filetransferutils==20.6
genie.libs.ops==20.6
genie.libs.parser==20.6
genie.libs.sdk==20.6
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
jsonpickle==1.4.1
junit-xml==1.9
MarkupSafe==1.1.1
multidict==4.7.6
netaddr==0.8.0
pathspec==0.8.0
ply==3.11
prettytable==0.7.2
psutil==5.7.2
pyasn1==0.4.8
pyats==20.6
pyats.aereport==20.6
pyats.aetest==20.6
pyats.async==20.6
pyats.connections==20.6
pyats.datastructures==20.6
pyats.easypy==20.6.1
pyats.kleenex==20.6
pyats.log==20.6.1
pyats.reporter==20.6
pyats.results==20.6
pyats.tcl==20.6
pyats.topology==20.6
pyats.utils==20.6
pycryptodomex==3.9.8
pysmi==0.3.4
pysnmp==4.4.12
python-engineio==3.13.1
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
tqdm==4.47.0
unicon==20.6
unicon.plugins==20.6
urllib3==1.25.9
xmltodict==0.12.0
yamllint==1.24.1
yarl==1.4.2
zipp==3.1.0
